RTDEControlInterface now has a threaded receive callback
It was necessary to implement a threaded receive callback function
for retrieving the data relevant to the control interface. This is
because of the RTDE data synchronization, which expects data to be
read continuously from the controller. If this is not the case the
controller will shutdown the connection after some time has passed,
resulting in the exception "connection reset by peer".

This fixes #10, but keep in mind that the data should be read
continuously from the controller and not with a large delay.